Refactorizando la API de Configuración
