+++ AguaSB.Configuracion.Tests/AguaSB.Configuracion.Tests.csproj
@@ -47,7 +47,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ConfiguracionTests.cs" />
+    <Compile Include="UtileriasDirectoriosTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
