+++ AguaSB.Configuracion.Tests/UtileriasDirectoriosTests.cs
@@ -4,7 +4,7 @@ using NUnit.Framework;
 namespace AguaSB.Configuracion.Tests
 {
     [TestFixture]
-    public class ConfiguracionTests
+    public class UtileriasDirectoriosTests
     {
         [TestCase("tmp", typeof(string), ".json", "tmp/String.json")]
         [TestCase("tmp/", typeof(string), ".json", "tmp/String.json")]
@@ -13,7 +13,7 @@ namespace AguaSB.Configuracion.Tests
         [TestCase("\\tmp/", typeof(string), ".json", "tmp/String.json")]
         public void Combinar_Deberia_CombinarCorrectamente(string subdirectorio, Type tipo, string extension, string resultado)
         {
-            var resultadoCombinacion = Configuracion.Combinar(subdirectorio, tipo.Name, extension);
+            var resultadoCombinacion = UtileriasDirectorios.Combinar(subdirectorio, tipo.Name, extension);
 
             Assert.AreEqual(resultado, resultadoCombinacion);
         }
